requirements.txt
datasets
python-dotenv
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.16
langchain-core==0.2.38
langchain-text-splitters==0.2.4
faiss-cpu

